# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TJ-T-Mart API",
        "description": "REST API for TJ-T-Mart — Dormitory Mini-Mart System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/produk": {
            "get": {
                "tags": [
                    "Produk",
                    "Public"
                ],
                "summary": "List active products with filters",
                "operationId": "070cf252f1e63a530eca818b332a72cb",
                "parameters": [
                    {
                        "name": "kategori_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by product name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mart_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated product list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/produk/{id}": {
            "get": {
                "tags": [
                    "Produk",
                    "Public"
                ],
                "summary": "Get product detail",
                "operationId": "5b0b36f23f31f36678f65352bbab36a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/produk": {
            "get": {
                "tags": [
                    "Admin",
                    "Produk"
                ],
                "summary": "Admin: list all products",
                "operationId": "92867e9cbafc2cd3a89dec6f063eea40",
                "responses": {
                    "200": {
                        "description": "Admin product list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Produk"
                ],
                "summary": "Admin: create product",
                "operationId": "85266bd5bdb78981d8ba93c41d53adef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nama_produk",
                                    "kategori_id",
                                    "harga",
                                    "stok"
                                ],
                                "properties": {
                                    "nama_produk": {
                                        "type": "string"
                                    },
                                    "kategori_id": {
                                        "type": "integer"
                                    },
                                    "harga": {
                                        "type": "number"
                                    },
                                    "stok": {
                                        "type": "integer"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "gambar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/produk/{id}": {
            "put": {
                "tags": [
                    "Admin",
                    "Produk"
                ],
                "summary": "Admin: update product",
                "operationId": "01d690be2ce690078194c323c9e41382",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nama_produk": {
                                        "type": "string"
                                    },
                                    "kategori_id": {
                                        "type": "integer"
                                    },
                                    "harga": {
                                        "type": "number"
                                    },
                                    "stok": {
                                        "type": "integer"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "gambar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin",
                    "Produk"
                ],
                "summary": "Admin: soft delete product",
                "operationId": "3ed56711968540e6bd6862c1e1ec5fe5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}